Create README.md
lhelloworld, This is Hephy from GitHub,
Just trying to see if this works I was not able to log from my Elive USB because of network/server not found. I'm  just testing. I logged in from f17 this is not my final hello world edited page.
I am so excited to learn to use gitHub and I can't wait to explore and use it to the maximum. Its been a long Journey since we first started tracing the history of Unix I wondered why I was asked by the teacher to trace the history of Unix when I was supposed to learn how Linux works. I asked my teacher why he wanted us to trace the history of Unix and He said it was necessary for me to know the History of Unix before I learn how Linux worked. It wasn't easy for me to trace the History of Unix since I had to research from the library. After researching the history of Unix I had a clear understanding how Unix is related to Linux. according to my research, I discovered that, Unix is Linux DNA. my second challenge was when the teacher asked us to build our own Linux and for me that was Impossible. "How was I going to do that?" I asked myself but after searching for information how to built my own Linux and after trying several times making mistakes, I was able to download Linux on Live USB. I was afraid to make mistakes at the beginning but after realizing that making mistakes is actually learning, I started exploring new skills and ideas. I was not afraid to make mistakes anymore. My 3rd challenge was when we decided as the whole F17 class to build our own distribution for our final exam. I was so excited and looking forward to learn more new skills. From the beginning it looked like a puzzle to me every week with the contributions from teams and team leads the puzzle was getting clear. I saw every student's contribution as part of the puzzle and I encouraged them to contribute because we needed every piece of the puzzle and not contributing was like losing parts of the puzzle. Each and every one of us had an important part to play in order to complete our puzzle. (Linux project) I felt honored to be the Project Lead and I knew right then that being a project Lead and trouble shooter much was required of me. I am so thankful for this wonderful opportunity because I have learned to be a leader and an example. I have learned leadership skills which I know will help me in my future endeavors. I learned to be effective, building relationships with each and every member of AbantOS Project, I have learned to be positive, focused, and encouraging. It was challenging at the beginning but when as we progressed, I learned to stand firm on my ground and take lead knowing that My teacher had faith in me when he picked me as The Project Lead. we are getting closer to complete our puzzle--AbantOS Project. I hope there will be more and bigger projects a head of us and this is just but, the beginning.
Welcome to the Enlightenment World.
Hephy